+++ IFSEngine/Animation/ControlPoint.cs
@@ -7,7 +7,7 @@ namespace IFSEngine.Animation
 {
     public class ControlPoint
     {
-        public float t;
+        public double t;
         public float Value;
         public Vector2 LeftTangent; //maybe angle enough
         public Vector2 RightTangent;
